+++ src/prob_any_missing.h
@@ -9,16 +9,15 @@ struct probAnyMissingFunctor
 {
     probAnyMissingFunctor() = default;
 
-    float operator()(const std::vector<float> &eventProbs, int numEvents);
+    double operator()(const std::vector<float> &eventProbs, int numEvents);
 
-    std::vector<float> vectorized(const std::vector<float> &eventProbs,
+    std::vector<double> vectorized(const std::vector<float> &eventProbs,
                                    unsigned int numEvents);
 
-    std::vector<float> vectorized(const std::vector<float> &eventProbs,
+    std::vector<double> vectorized(const std::vector<float> &eventProbs,
                                    unsigned int minNumEvents,
                                    unsigned int maxNumEvents);
 
-    // std::vector<float> baseVec{};
     CombinationIndicesGenerator c;
 };
